+++ tests/Support/NamespacedListRequest.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace BYanelli\Roma\Tests\Support;
+
+readonly class NamespacedListRequest
+{
+    /** @var list<NamespacedItem> */
+    public array $items;
+}